requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
country-currencies==0.2
Django==3.2.4
django-mssql-backend==2.8.1
djangorestframework==3.12.4
geoip2==4.2.0
idna==2.10
marshmallow==2.15.6
maxminddb==2.0.3
multidict==5.1.0
openexchangerates==0.1.1
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
transbank-sdk==1.5.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
